Error happens in the Crop Image
It crop the image in that way the image end side also being crop and not get the full and clear image of the hand. So now lets work on that so we can increase the width and height frame of the crop image. So we can get the clear image